+++ src/app/api/cities/search/route.ts
@@ -1,5 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import { getCities } from "../index";
+import { SEARCH_SUGGESTION_LIMIT } from "../../../utils/config";
 
 export async function GET(request: NextRequest) {
   const { searchParams } = new URL(request.url);
@@ -8,7 +9,7 @@ export async function GET(request: NextRequest) {
   const cities = await getCities();
   const filteredCities = cities
     .filter((city: any) => city.name.toLowerCase().includes(q))
-    .slice(0, 5);
+    .slice(0, SEARCH_SUGGESTION_LIMIT);
 
   return NextResponse.json(filteredCities);
 }
