+++ src/app/utils/config.ts
@@ -1,2 +1,6 @@
 export const API_URL = process.env.NEXT_PUBLIC_OPENWEATHER_API_URL!;
 export const API_KEY = process.env.NEXT_PUBLIC_OPENWEATHER_API_KEY!;
+
+export const TOAST_DURATION = 3000;
+export const SEARCH_SUGGESTION_LIMIT = 5;
+export const TABLE_ITEM_PER_PAGE = 10;
